chore: making compiling stricter, move to C++20

+++ factory/calorimeter/ODDPolyhedraBarrelCalorimeter_geo.cpp
@@ -69,7 +69,7 @@ static Ref_t create_detector(Detector& description, xml_h e, SensitiveDetector s
   double layerInnerAngle = (M_PI / 2 - layerOuterAngle);
   double layer_pos_z = -(staveThickness / 2);
   double layer_dim_x = innerFaceLen / 2 - gap * 2;
-  int layer_num = 1;
+  size_t layer_num = 1;
 
   // Copy of the LayeredCaloData from CLD factory
   // Create caloData object to extend driver with data required for reconstruction
@@ -109,9 +109,9 @@ static Ref_t create_detector(Detector& description, xml_h e, SensitiveDetector s
 
     // Loop over repeats for this layer.
     for (int j = 0; j < repeat; j++) {
-      string layer_name = _toString(layer_num, "layer%d");
+      string layer_name = _toString(static_cast<int>(layer_num), "layer%d");
       double layer_thickness = lay->thickness();
-      DetElement layer(stave, layer_name, layer_num);
+      DetElement layer(stave, layer_name, static_cast<int>(layer_num));
 
       // Layer position in Z within the stave.
       layer_pos_z += layer_thickness / 2;
@@ -174,13 +174,13 @@ static Ref_t create_detector(Detector& description, xml_h e, SensitiveDetector s
 
       // Layer physical volume.
       PlacedVolume layer_phv = staveInnerVol.placeVolume(layer_vol, Position(0, 0, layer_pos_z));
-      layer_phv.addPhysVolID("layer", layer_num);
+      layer_phv.addPhysVolID("layer", static_cast<int>(layer_num));
       layer.setPlacement(layer_phv);
 
       // The rest of the data is constant; only the distance needs to be updated
       // Store the position up to the inner face of the layer
       caloLayer.distance = rmin + layer_pos_z + staveThickness / 2 - layer_thickness / 2;
-      std::cout << "Layer: " << layer_num << " Rmin: " << rmin << " layer_pos_z: " << layer_pos_z
+      std::cout << "Layer: " << static_cast<int>(layer_num) << " Rmin: " << rmin << " layer_pos_z: " << layer_pos_z
                 << " Dist: " << caloLayer.distance
                 << " inner_thickness: " << caloLayer.inner_thickness
                 << " outer_thickness: " << caloLayer.outer_thickness
